refactor: 피드 조회시에 스터디 이름도 반환
