+++ backend/src/test/java/com/yigongil/backend/acceptance/steps/FeedSteps.java
@@ -7,12 +7,11 @@ import static org.junit.jupiter.api.Assertions.assertAll;
 import com.yigongil.backend.request.CertificationCreateRequest;
 import com.yigongil.backend.request.FeedPostCreateRequest;
 import com.yigongil.backend.response.CertificationResponse;
-import com.yigongil.backend.response.FeedPostResponse;
+import com.yigongil.backend.response.feed.FeedPageResponse;
 import io.cucumber.java.en.Then;
 import io.cucumber.java.en.When;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import java.util.List;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
@@ -76,12 +75,12 @@ public class FeedSteps {
                                                         .all()
                                                         .statusCode(HttpStatus.OK.value())
                                                         .extract();
-        List<FeedPostResponse> responses = response.jsonPath().getList(".", FeedPostResponse.class);
+        FeedPageResponse feedPageResponse = response.as(FeedPageResponse.class);
 
         assertAll(
-                () -> assertThat(responses).hasSize(1),
-                () -> assertThat(responses.get(0).author().nickname()).isEqualTo(author),
-                () -> assertThat(responses.get(0).content()).isEqualTo(content)
+                () -> assertThat(feedPageResponse.feeds()).hasSize(1),
+                () -> assertThat(feedPageResponse.feeds().get(0).author().nickname()).isEqualTo(author),
+                () -> assertThat(feedPageResponse.feeds().get(0).content()).isEqualTo(content)
         );
     }
